requirements.txt
opencv-python==4.12.0.88
pytesseract==0.3.9
passporteye==2.2.2
Pillow==10.4.0
Flask==2.0.3
PyYAML==6.0.1
Werkzeug==2.0.3
